requirements.txt
Pillow
os
numpy
opencv-python